fix: complex to polar makes sense
As pointed out in Issue #423, the `complex_to_polar` function did not make sense. Now it gives a complex number in component form, for a student to convert into polar form.

This is not a perfect function, but at least it is usable now. In future, we may want to have the theta be chosen from amongst a set of multiples/fractions of pi, that tend to make for nicer questions and answers.